==> Pandas/1. Pandas - Loading Data FroM Collection and File.ipynb <==
import numpy as np
import pandas as pd

# The core data abstraction layer in Pandas is called a DataFrame
# Any data that you load/initialize using Pandas will be represented in the form of DataFrame
# To create a Dataframe, you can use the following two approach
# 1. Create DF using collection object (list)
# 2. Create DF by loading a file

#Example using list

list1 = [[1,'Mohit',1000], [2,'Arun',2000]]

empDataFrameFromList = pd.DataFrame(list1)
empDataFrameFromList

# DataFrame data is represented using Row and Column indexes
# You can replace Column indexes with column names

empDataFrameFromList.columns = ['eid','ename','esal']
empDataFrameFromList

#Variation 2:
list2 = [[1,'Mohit',5000], [2,'Arun',8000], [3,'Aman',9899]]

empDataFrameFromList2 = pd.DataFrame(list2 , columns=['empid','ename','esal'])
empDataFrameFromList2

# Create DataFrame by File
# LOad a delimited file --> Any file which has a column seperator character.
# e.g. CSV (Comma Seperated Values), TSV (Tab Seperated Values) etc.

employeeDataFrame = pd.read_csv("employee.csv", header=None, names=['eid','ename','esal'])
employeeDataFrame

#?pd.read_csv

empWithHeaderDF = pd.read_csv("employeeWithHeaders.csv", sep=',')
empWithHeaderDF

#Load the data from remote path 
irisFromRemote = pd.read_csv("https://raw.githubusercontent.com/uiuc-cse/data-fa14/gh-pages/data/iris.csv")
irisFromRemote.head(5)

irisFromRemote.tail(5)
