==> Pandas/2. Series vs DataFrame.ipynb <==
import numpy as np
import pandas as pd

empWithHeaderDF = pd.read_csv("employeeWithHeaders.csv", sep=',')
empWithHeaderDF

#To get the information of the dataframe
empWithHeaderDF.info()

type(empWithHeaderDF)

# Series in Pandas
# Seriees is a column.Very similar to Numpy Array with dimension (X,1)
# A set of Series forms a DataFrame

empWithHeaderDF.esal

type(empWithHeaderDF.esal)

# You can convert the DataFrame/Series into a Numpy Array

empArray = empWithHeaderDF.values #.values convert DF into Numpy Array
print(type(empArray))
empArray

# There will be a situation where you will need to convert DF into Numpy

# Situation 1:

# If you are writing a code to create Statistical Model (Machine Learning Model/Algo) in Python
# the package that we use is Sci-kit learn(SKLEARN). SKLEARN expects your data to be in the form
# of Numpy Array.

# Situation 2:

# If you want to perform Statistical Testing (Hypothesis Testing) then the formula expects your
# data to be either in the form of Matrix or in the form of ndarray (Numpy)

# Situation 3:

# To implement Deep Learning Neural Networks your data must be represented in the form of Numpy


# Your plan of action will be
# 1. Load the data from file using Pandas
# 2. Ensure your dataset is in compliance with the next process( Statistical Modelling, Statistical Testing, NN creation)
# 3. Convert the dataframe into Numpy Array
# 4. Implement  Statistical Modelling, Statistical Testing, NN creation

# Even reverse is possible
pd.DataFrame(empArray , columns=['eid','ename','esal'])
