==> Pandas/3. Operations in Series.ipynb <==
import numpy as np 
import pandas as pd

empWithHeaderDF = pd.read_csv("employeeWithHeaders.csv", sep=',')
empWithHeaderDF

#Use-case: I want to calculate the yearlySalary of the employee and add a new column called yearlySalary

yearlySalary = empWithHeaderDF.esal * 12
yearlySalary

type(yearlySalary)

#Pandas Dataframes are mutable in nature

empWithHeaderDF['yearlySalary'] = yearlySalary
empWithHeaderDF

# Add a new column using a collection object

deptList = ['HR','Ops','','','Ops','HR']

# Convert the collection object into series
# Technique -- Load the collection as series

deptSeries = pd.Series(deptList)

#Add Series in dataFrame
empWithHeaderDF['department'] = deptSeries
empWithHeaderDF

# Use-case: I want to create a new column called UpdatedYearlySalary based on the following increment/hike conditions
# If the salary is less than equal to 1500, then inc the sal by 10%
# If the salary is between 1501 and 10000, then inc the sal by 5%
# If the salary is between 10001 and 20000, then inc the sal by 2.5%
# If the salary is greater than 20000 , then no inc

# Pandas recommends to create a Function that can implement the logic
# and use apply method to implement function in a series

def incrementSalary(salary):
    newSal = 0
    
    if salary <= 1500:
        newSal = salary * 1.1
    elif salary <= 10000:
        newSal = salary * 1.05
    elif salary <= 20000:
        newSal = salary * 1.025
    else:
        newSal = salary
    
    return newSal

empWithHeaderDF['UpdatedYearlySalary'] = empWithHeaderDF.esal.apply(incrementSalary) * 12
empWithHeaderDF

empWithHeaderDF.department = pd.Series(['HR','Ops','Admin','Admin','Ops','HR'])
empWithHeaderDF

#Store the dataframe into a file
empWithHeaderDF.to_csv('FinalOutput.csv' , index=False)
