==> Pandas/4. Traversing the DataFrame.ipynb <==
import numpy as np
import pandas as pd

datasetExample = pd.read_csv('FinalOutput.csv')

datasetExample

# first 3 records 
# ename,dept,UpdatedSalary
datasetExample.iloc[ 0:3 , [1,4,5]]

# first 3 records 
# UpdatedSalary,ename,dept
datasetExample.iloc[ 0:3 , [5,1,4]]

# Data Traversal
# Type1 -- Traversing the data using Index
#
# iloc --> indexLocation
# iloc[rowIndex/rowRange/rowList , columnIndex/columnRange/columnList]


datasetExample.iloc[ 0:3 , :]

datasetExample.iloc[ 0:3 , 0:4]

datasetExample.iloc[ 0:3 , [0,4] ]
