==> Pandas/5. DataFrame Operations.ipynb <==
import numpy as np
import pandas as pd

datasetExample = pd.read_csv('FinalOutput.csv')

### Operation 1:  Use of Relational,logical and comparison oeprations

#1. Extract the records of employees whose esal is greater than 5000

datasetExample[ datasetExample.esal > 5000 ]

#2. Extract only ename and department of employees whose esal is greater than 5000
datasetExample[ datasetExample.esal > 5000 ].iloc[:,[1,4]]

#3. Extract those records whose sal is greater than 6500 and belongs to Ops dept
datasetExample[ (datasetExample.esal > 6500) & (datasetExample.department == 'Ops') ]

#4. Display the name of employees whose salary is greater than 6500 and belongs to HR dept
datasetExample[(datasetExample.esal > 6500) & (datasetExample.department == 'HR')].ename

datasetExample[(datasetExample.esal > 6500) & (datasetExample.department == 'HR')].iloc[:,[1]]

datasetExample

# Replace the salary of the employee as 7000 if the current salary is less than 1500
# assignment

### Operation 2 - Dealing with Duplicate Rows

datasetExample

datasetExample.drop_duplicates(inplace= True) #Designed only for duplicate rows

datasetExample

# The major demerit of DropDuplicate is the index is never reset automatically which may impact the
# fetch cycle of the data when performing EDA or Statistical Modelling.
# So the best practice is to reset the index as soon as DropDuplicate is performed
# U will see next Week ;)


### Operation 3: Performing  Merge Operations iN Pandas

dfExample1 = pd.DataFrame([[4,'QA'],[1,'HR'],[3,'Dev'],[2,'Ops']] , columns=['eid','dept'])

dfExample1

dfExample2 = pd.DataFrame([[1,'Prashant'],[2,'Gokul'],[3,'Guna']] , columns=['eid','ename'])

dfExample2

# Use-case. I want to perform Join operation of two dfs based on eid
# We will use merge
resultDF = pd.merge(dfExample2,dfExample1)
resultDF
# The merge worked in this case because both dataframe have the common column eid
# that too with the same name

dept = pd.DataFrame([[4,'QA'],[1,'HR'],[3,'Dev'],[2,'Ops']] , columns=['eid','dept'])
emp = pd.DataFrame([[1,'Prashant'],[2,'Gokul'],[3,'Guna']] , columns=['empid','ename'])

dept

emp

# Use-case. I want to perform Join operation of two dfs based on empid and eid resp.
resultDF2 = pd.merge(emp,dept ,left_on='empid', right_on='eid' )
resultDF2

finalAnswer = resultDF2.iloc[:,[0,1,3]]

finalAnswer

resultDF2 = pd.merge(emp,dept ,left_on='empid', right_on='eid')
resultDF2

### Operation 4 - Concat Operation

empExample = pd.DataFrame([[1,'Prashant'],[2,'Gokul'],[3,'Guna']] , columns=['empid','ename'])

empExample2 = pd.DataFrame([[4,'Nik'],[5,'Ashish'],[6,'Asha']] , columns=['empid','ename'])

empExample

empExample2

#Rowwise Concatenation --- Ensure the column names are same in all DFs.
resultEmp = pd.concat( [empExample,empExample2] , axis = 0)
resultEmp

empExample3 = pd.DataFrame([[4,'Nik'],[5,'Ashish'],[6,'Asha']] , columns=['empid','empname'])
empExample3.columns = ['empid','ename']
resultEmp2 = pd.concat( [empExample,empExample3] , axis = 0)
resultEmp2

#Columnwise Concat
empExample = pd.DataFrame([[1,'Prashant'],[2,'Gokul'],[3,'Guna']] , columns=['empid','ename'])
dept = pd.DataFrame([['Mumbai','QA'],['Chennai','HR'],['Bangalore','Dev']] , columns=['ecity','dept'])


empExample

dept

#Columnwise Concatenation --- Ensure the Row Indexes are same in all DFs.
resultEmp3 = pd.concat( [empExample,dept] , axis = 1)
resultEmp3

### Operation 6 --- Deletion 

datasetExample

#Column Deletion (axis = 1) ---> UpdatedYearlySalary

datasetExample.drop(['UpdatedSalary','UpdatedYearlySalary'] , axis = 1 , inplace=True)


?datasetExample.drop

datasetExample

#Row Deletion (axis=0)
datasetExample.drop([5] , axis=0, inplace=True)

datasetExample

datasetExample.drop( datasetExample[datasetExample.eid == 3].index , axis=0)

datasetExample.iloc[:,3:6].index
